Add type to variable in oldest_first.jl

+++ src/strategies/oldest_first.jl
@@ -39,7 +39,7 @@ function run_oldest_first(
 	push!(utility_A.benefits, 1.0)
 	push!(utility_B.benefits, 1.0)
 
-	idx_strategy = 3
+	idx_strategy::Int64 = 3
 	VaxCounselors.write_benefits_CSV(
 		utility_A.benefits,
 		utility_B.benefits,
